squash: move some logic into base
